==> notebooks/02_data_cleaning.ipynb <==
# vct_2024/
#     agents/
#         agents_pick_rates.csv
#     ids/
#         players_ids.csv
#         teams_ids.csv
#         tournaments_stages_match_types_ids.csv
#         tournaments_stages_matches_games_ids.csv
#     matches/
#         maps_played.csv
#         maps_scores.csv
#         overview.csv
#         scores.csv
#     players_stats/
#         players_stats.csv

import pandas as pd
from pathlib import Path

df = pd.read_parquet("../data/raw/vct_2024/players_stats/players_stats.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Player', 'Teams', 'Agents',
#        'Rounds Played', 'Rating', 'Average Combat Score', 'Kills:Deaths',
#        'Kill, Assist, Trade, Survive %', 'Average Damage Per Round',
#        'Kills Per Round', 'Assists Per Round', 'First Kills Per Round',
#        'First Deaths Per Round', 'Headshot %', 'Clutch Success %',
#        'Clutches (won/played)', 'Maximum Kills in a Single Map', 'Kills',
#        'Deaths', 'Assists', 'First Kills', 'First Deaths'],
#       dtype='str')

df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Player" : "player_name",
    "Teams" : "team_name",
    "Agents" : "agent_name",
    "Rounds Played" : "rounds_played",
    "Rating" : "rating",
    "Average Combat Score" : "average_combat_score",
    "Kills:Deaths" : "kills_deaths_ratio",
    "Kill, Assist, Trade, Survive %" : "kill_assist_trade_survive_percentage",
    "Average Damage Per Round" : "average_damage_per_round",
    "Kills Per Round" : "kills_per_round",
    "Assists Per Round" : "assists_per_round",
    "First Kills Per Round" : "first_kills_per_round",
    "First Deaths Per Round" : "first_deaths_per_round",
    "Headshot %" : "headshot_percentage",
    "Clutch Success %" : "clutch_success_percentage",
    "Clutches (won/played)" : "clutches_won_played",
    "Maximum Kills in a Single Map" : "maximum_kills_single_map",
    "Kills" : "kills",
    "Deaths" : "deaths",
    "Assists" : "assists",
    "First Kills" : "first_kills",
    "First Deaths" : "first_deaths"
})

df.columns

df['kill_assist_trade_survive_percentage'] = df['kill_assist_trade_survive_percentage'].str.rstrip('%').astype('float') / 100.0
df['headshot_percentage'] = df['headshot_percentage'].str.rstrip('%').astype('float') / 100.0
df['clutch_success_percentage'] = df['clutch_success_percentage'].str.rstrip('%').astype('float') / 100.0

# handle nan values in 'clutches_won_played' before splitting
df['clutches_won_played'] = df['clutches_won_played'].fillna('0/0')
df[['clutches_won', 'clutches_played']] = df['clutches_won_played'].str.split('/', expand=True).astype('int')
df = df.drop(columns=['clutches_won_played'])

df['clutch_success_percentage'] = df['clutch_success_percentage'].fillna(0.0)

df.info()

path = Path("../data/clean/vct_2024/players_stats")
path.mkdir(parents=True, exist_ok=True)


df.to_parquet(path / "players_stats.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/agents/agents_pick_rates.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Map', 'Agent', 'Pick Rate'], dtype='str')
df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Map" : "map_name",
    "Agent" : "agent_name",
    "Pick Rate" : "pick_rate"
})

df.columns

df['pick_rate'] = df['pick_rate'].str.rstrip('%').astype('float') / 100.0

df.info()

path = Path("../data/clean/vct_2024/agents")
path.mkdir(parents=True, exist_ok=True)

df.to_parquet(path / "agents_pick_rates.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/matches/maps_played.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Match Name', 'Map'], dtype='str')
df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Match Name" : "match_name",
    "Map" : "map_name"
})

path = Path("../data/clean/vct_2024/matches")
path.mkdir(parents=True, exist_ok=True)

df.to_parquet(path / "maps_played.parquet", index=False)

# vct_2024/
#     agents/
#         agents_pick_rates.csv-------------------
#     ids/
#         players_ids.csv           
#         teams_ids.csv
#         tournaments_stages_match_types_ids.csv
#         tournaments_stages_matches_games_ids.csv
#     matches/
#         maps_played.csv------------------
#         maps_scores.csv------------------
#         overview.csv---------------------
#         scores.csv-----------------------
#     players_stats/
#         players_stats.csv-----------------

df = pd.read_parquet("../data/raw/vct_2024/matches/maps_scores.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Match Name', 'Map', 'Team A',
#        'Team A Score', 'Team A Attacker Score', 'Team A Defender Score',
#        'Team A Overtime Score', 'Team B', 'Team B Score',
#        'Team B Attacker Score', 'Team B Defender Score',
#        'Team B Overtime Score', 'Duration'],
#       dtype='str')

df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Match Name" : "match_name",
    "Map" : "map_name",
    "Team A" : "team_a_name",
    "Team A Score" : "team_a_score",
    "Team A Attacker Score" : "team_a_attacker_score",
    "Team A Defender Score" : "team_a_defender_score",
    "Team A Overtime Score" : "team_a_overtime_score",
    "Team B" : "team_b_name",
    "Team B Score" : "team_b_score",
    "Team B Attacker Score" : "team_b_attacker_score",
    "Team B Defender Score" : "team_b_defender_score",
    "Team B Overtime Score" : "team_b_overtime_score",
    "Duration" : "duration"
})

df.info()

df['duration'] = df['duration'].apply(
    lambda x: "00:" + x if isinstance(x, str) and x.count(":") == 1 else x
)

df['duration'] = pd.to_timedelta(df['duration'])

df['duration_seconds'] = df['duration'].dt.total_seconds()
df = df.drop(columns=['duration'])

df.to_parquet(path / "maps_scores.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/matches/overview.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Match Name', 'Map', 'Player',
#        'Team', 'Agents', 'Rating', 'Average Combat Score', 'Kills', 'Deaths',
#        'Assists', 'Kills - Deaths (KD)', 'Kill, Assist, Trade, Survive %',
#        'Average Damage Per Round', 'Headshot %', 'First Kills', 'First Deaths',
#        'Kills - Deaths (FKD)', 'Side'],
#       dtype='str')

df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Match Name" : "match_name",
    "Map" : "map_name",
    "Player" : "player_name",
    "Team" : "team_name",
    "Agents" : "agent_name",
    "Rating" : "rating",
    "Average Combat Score" : "average_combat_score",
    "Kills" : "kills",
    "Deaths" : "deaths",
    "Assists" : "assists",
    "Kills - Deaths (KD)" : "kills_deaths_ratio",
    "Kill, Assist, Trade, Survive %" : "kill_assist_trade_survive_percentage",
    "Average Damage Per Round" : "average_damage_per_round",
    "Headshot %" : "headshot_percentage",
    "First Kills" : "first_kills",
    "First Deaths" : "first_deaths",
    "Kills - Deaths (FKD)" : "first_kills_deaths_ratio",
    "Side" : "side"
})

df.info()

df['kill_assist_trade_survive_percentage'] = df['kill_assist_trade_survive_percentage'].str.rstrip('%').astype('float') / 100.0
df['headshot_percentage'] = df['headshot_percentage'].str.rstrip('%').astype('float') / 100.0

# Side has values attack, defence, both. Convert them to lowercase for consistency, then to pandas category type
df['side'] = df['side'].str.lower().astype('category')

df.to_parquet(path / "overview.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/matches/scores.parquet")

df.columns

# Index(['Tournament', 'Stage', 'Match Type', 'Match Name', 'Team A', 'Team B',
#        'Team A Score', 'Team B Score', 'Match Result'],
#       dtype='str')

df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Stage" : "stage_name",
    "Match Type" : "match_type_name",
    "Match Name" : "match_name",
    "Team A" : "team_a_name",
    "Team B" : "team_b_name",
    "Team A Score" : "team_a_score",
    "Team B Score" : "team_b_score",
    "Match Result" : "match_result"
})

df['winning_team_name'] = df['match_result'].str.replace(' won', '', regex=False)
df = df.drop(columns=['match_result'])

df.to_parquet(path / "scores.parquet", index=False)

# vct_2024/
#     agents/
#         agents_pick_rates.csv-------------------
#     ids/
#         players_ids.csv-----------------------    
#         teams_ids.csv------------------------
#         tournaments_stages_match_types_ids.csv
#         tournaments_stages_matches_games_ids.csv
#     matches/
#         maps_played.csv------------------
#         maps_scores.csv------------------
#         overview.csv---------------------
#         scores.csv-----------------------
#     players_stats/
#         players_stats.csv-----------------

df = pd.read_parquet("../data/raw/vct_2024/ids/players_ids.parquet")

df.columns

# Index(['Player', 'Player ID'], dtype='str')
df = df.rename(columns={
    "Player" : "player_name",
    "Player ID" : "player_id"
})

path = Path("../data/clean/vct_2024/ids")
path.mkdir(parents=True, exist_ok=True)

df.to_parquet(path / "players_ids.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/ids/teams_ids.parquet")

df.columns

# Index(['Team', 'Team ID'], dtype='str')
df = df.rename(columns={
    "Team" : "team_name",
    "Team ID" : "team_id"
})

df.to_parquet(path / "teams_ids.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/ids/tournaments_stages_match_types_ids.parquet")

df.columns

# Index(['Tournament', 'Tournament ID', 'Stage', 'Stage ID', 'Match Type',
#        'Match Type ID'],
#       dtype='str')
df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Tournament ID" : "tournament_id",
    "Stage" : "stage_name",
    "Stage ID" : "stage_id",
    "Match Type" : "match_type_name",
    "Match Type ID" : "match_type_id"
})

df = df.dropna(subset=['stage_id', 'match_type_id'])

# convert stage_id and match_type_id to integers
df['stage_id'] = df['stage_id'].astype('int')
df['match_type_id'] = df['match_type_id'].astype('int')

df.to_parquet(path / "tournaments_stages_match_types_ids.parquet", index=False)

df = pd.read_parquet("../data/raw/vct_2024/ids/tournaments_stages_matches_games_ids.parquet")

df.columns

# Index(['Tournament', 'Tournament ID', 'Stage', 'Stage ID', 'Match Type',
#        'Match Name', 'Match ID', 'Map', 'Game ID'],
#       dtype='str')
df = df.rename(columns={
    "Tournament" : "tournament_name",
    "Tournament ID" : "tournament_id",
    "Stage" : "stage_name",
    "Stage ID" : "stage_id",
    "Match Type" : "match_type_name",
    "Match Name" : "match_name",
    "Match ID" : "match_id",
    "Map" : "map_name",
    "Game ID" : "game_id"
})

df.to_parquet(path / "tournaments_stages_matches_games_ids.parquet", index=False)
