==> notebooks/03_create_dimentions.ipynb <==
import pandas as pd
from pathlib import Path

# Paths
BASE = Path("../data/clean/vct_2024")
OUT = Path("../data/gold/dimensions")

OUT.mkdir(parents=True, exist_ok=True)

players_raw = pd.read_parquet(BASE / "ids/players_ids.parquet")

players_raw.head()

dim_players = players_raw[["player_id", "player_name"]].drop_duplicates().sort_values("player_id").reset_index(drop=True)

output_path = OUT / "dim_player.parquet"

dim_players.to_parquet(output_path, index=False)

teams_raw = pd.read_parquet(BASE / "ids/teams_ids.parquet")

teams_raw.head()

dim_teams = teams_raw[["team_id", "team_name"]].drop_duplicates().sort_values("team_id").reset_index(drop=True)

output_path = OUT / "dim_team.parquet"

dim_teams.to_parquet(output_path, index=False)

t_raw = pd.read_parquet(BASE / "ids/tournaments_stages_match_types_ids.parquet")

t_raw.head()

dim_tournament = (
    t_raw[
        [
            "tournament_id",
            "tournament_name",
            "stage_id",
            "stage_name",
            "match_type_id",
            "match_type_name",
        ]
    ]
    .drop_duplicates()
    .sort_values(
        ["tournament_id", "stage_id", "match_type_id"]
    )
    .reset_index(drop=True)
)

dim_tournament.head()

dim_tournament.to_parquet(
    OUT / "dim_tournament.parquet",
    index=False
)

valorant_maps = [
    "Ascent",
    "Bind",
    "Breeze",
    "Corrode",
    "Fracture",
    "Haven",
    "Icebox",
    "Lotus",
    "Pearl",
    "Split",
    "Sunset",
    "Abyss"
]


valorant_agents = [
    "Astra",
    "Breach",
    "Brimstone",
    "Chamber",
    "Clove",
    "Cypher",
    "Deadlock",
    "Fade",
    "Gekko",
    "Harbor",
    "Iso",
    "Jett",
    "KayO",
    "Killjoy",
    "Neon",
    "Omen",
    "Phoenix",
    "Raze",
    "Reyna",
    "Sage",
    "Skye",
    "Sova",
    "Tejo",
    "Viper",
    "Veto",
    "Vyse",
    "Waylay",
    "Yoru"
]

valorant_agents = [agent.lower() for agent in valorant_agents]

dim_map = pd.DataFrame({
    "map_name": sorted(valorant_maps)
})

dim_map["map_id"] = dim_map.index + 1

dim_map = dim_map[["map_id", "map_name"]]

dim_map_path = OUT / "dim_map.parquet"

dim_map.to_parquet(dim_map_path, index=False)

dim_agent = pd.DataFrame({
    "agent_name": sorted(valorant_agents)
})

dim_agent["agent_id"] = dim_agent.index + 1

dim_agent = dim_agent[["agent_id", "agent_name"]]

dim_agent_path = OUT / "dim_agent.parquet"

dim_agent.to_parquet(dim_agent_path, index=False)

m_raw = pd.read_parquet(BASE / "ids/tournaments_stages_matches_games_ids.parquet")
t_raw = pd.read_parquet(BASE / "ids/tournaments_stages_match_types_ids.parquet")

m_raw.head()

m_merged = m_raw.merge(
    t_raw[
        [
            "tournament_name",
            "stage_name",
            "match_type_name",
            "match_type_id",
        ]
    ],
    on=[
        "tournament_name",
        "stage_name",
        "match_type_name",
    ],
    how="left"
)


m_merged.columns

dim_match = (
    m_merged[
        [
            "match_id",
            "game_id",
            "match_name",
            "map_name",
            "tournament_id",
            "stage_id",
            "match_type_id",
        ]
    ]
    .drop_duplicates()
    .sort_values(["match_id", "game_id"])
    .reset_index(drop=True)
)


dim_match.head()

dim_match_path = OUT / "dim_match.parquet"

dim_match.to_parquet(dim_match_path, index=False)
