==> notebooks/04_create_overview_fact.ipynb <==
import pandas as pd
from pathlib import Path

# Paths
BASE = Path("../data/clean/vct_2024")
OUT = Path("../data/gold/facts")
DIM = Path("../data/gold/dimensions")

OUT.mkdir(parents=True, exist_ok=True)

# Silver source
overview = pd.read_parquet(
    BASE / "matches/overview.parquet"
)

# Dimensions
dim_player = pd.read_parquet(DIM / "dim_player.parquet")
dim_team = pd.read_parquet(DIM / "dim_team.parquet")
dim_tournament = pd.read_parquet(DIM / "dim_tournament.parquet")
dim_match = pd.read_parquet(DIM / "dim_match.parquet")
dim_map = pd.read_parquet(DIM / "dim_map.parquet")
dim_agent = pd.read_parquet(DIM / "dim_agent.parquet")

# Remove aggregate rows
overview = overview[
    overview["map_name"] != "All Maps"
].copy()

overview.head()

fact = overview.merge(
    dim_player,
    on="player_name",
    how="left"
)

print(fact['player_id'].isna().sum())

fact = fact.merge(
    dim_team,
    on="team_name",
    how="left"
)

print(fact['team_id'].isna().sum())

fact = fact.merge(
    dim_tournament,
    on=[
        "tournament_name",
        "stage_name",
        "match_type_name"
    ],
    how="left"
)


print(fact["tournament_id"].isna().sum())

fact.columns

dim_match.columns

fact = fact.merge(
    dim_match,
    on=[
        "tournament_id",
        "stage_id",
        "match_type_id",
        "match_name",
        "map_name"
    ],
    how="left"
)

print(fact["match_id"].isna().sum())

fact = fact.merge(
    dim_map,
    on="map_name",
    how="left"
)

print(fact["map_id"].isna().sum())

fact = fact.merge(
    dim_agent,
    on="agent_name",
    how="left"
)

print(fact["agent_id"].isna().sum())

fact.columns

# Index(['tournament_name', 'stage_name', 'match_type_name', 'match_name',
#        'map_name', 'player_name', 'team_name', 'agent_name', 'rating',
#        'average_combat_score', 'kills', 'deaths', 'assists',
#        'kills_deaths_ratio', 'kill_assist_trade_survive_percentage',
#        'average_damage_per_round', 'headshot_percentage', 'first_kills',
#        'first_deaths', 'first_kills_deaths_ratio', 'side', 'player_id',
#        'team_id', 'tournament_id', 'stage_id', 'match_type_id', 'match_id',
#        'game_id', 'map_id', 'agent_id'],
#       dtype='str')

fact_player_match_map_stats = fact[
    [
        # Keys
        "match_id",
        "game_id",

        "player_id",
        "team_id",

        "tournament_id",
        "stage_id",
        "match_type_id",

        "map_id",
        "agent_id",

        "side",

        # Metrics
        "rating",
        "average_combat_score",
        "kills",
        "deaths",
        "assists",
        "kills_deaths_ratio",
        "kill_assist_trade_survive_percentage",
        "average_damage_per_round",
        "headshot_percentage",
        "first_kills",
        "first_deaths",
        "first_kills_deaths_ratio"
    ]
].copy()


fact_player_match_map_stats.info()

fact_player_match_map_stats.isna().sum()

fact_player_match_map_stats

out_path = OUT / "fact_player_match_map_stats.parquet"

fact_player_match_map_stats.to_parquet(
    out_path,
    index=False
)
