==> notebooks/05_create_scores_fact.ipynb <==
import pandas as pd
from pathlib import Path

BASE = Path("../data/clean/vct_2024")
DIM = Path("../data/gold/dimensions")
OUT = Path("../data/gold/facts")

OUT.mkdir(parents=True, exist_ok=True)

# Silver source
scores = pd.read_parquet(
    BASE / "matches/scores.parquet"
)

# Dimensions
dim_team = pd.read_parquet(DIM / "dim_team.parquet")
dim_match = pd.read_parquet(DIM / "dim_match.parquet")
dim_tournament = pd.read_parquet(DIM / "dim_tournament.parquet")

fact = scores.merge(
    dim_team,
    left_on="team_a_name",
    right_on="team_name",
    how="left"
).rename(
    columns={"team_id": "team_a_id"}
)

fact = fact.drop(columns=["team_name"])

print(fact["team_a_id"].isna().sum())

fact = fact.merge(
    dim_team,
    left_on="team_b_name",
    right_on="team_name",
    how="left"
).rename(
    columns={"team_id": "team_b_id"}
)

fact = fact.drop(columns=["team_name"])

print(fact['team_b_id'].isna().sum())

fact = fact.merge(
    dim_team,
    left_on="winning_team_name",
    right_on="team_name",
    how="left"
).rename(
    columns={"team_id": "winning_team_id"}
)

fact = fact.drop(columns=["team_name"])

print(fact["winning_team_id"].isna().sum())

fact = fact.merge(
    dim_tournament,
    on=[
        "tournament_name",
        "stage_name",
        "match_type_name"
    ],
    how="left"
)

print(fact["tournament_id"].isna().sum())

dim_match

fact = fact.merge(
    dim_match[
        [
            "match_id",
            "match_name",
            "tournament_id",
            "stage_id",
            "match_type_id"
        ]
    ].drop_duplicates(),
    on=[
        "tournament_id",
        "stage_id",
        "match_type_id",
        "match_name"
    ],
    how="left"
)

print(fact["match_id"].isna().sum())

fact_match_results = fact[
    [
        # Keys
        "match_id",

        "tournament_id",
        "stage_id",
        "match_type_id",

        "team_a_id",
        "team_b_id",
        "winning_team_id",

        # Scores
        "team_a_score",
        "team_b_score",
    ]
].copy()

print(fact_match_results.duplicated("match_id").sum())

fact_match_results.isna().sum()

out_path = OUT / "fact_match_results.parquet"

fact_match_results.to_parquet(
    out_path,
    index=False
)
