==> notebooks/06_create_players_stats_fact.ipynb <==
import pandas as pd
from pathlib import Path

BASE = Path("../data/clean/vct_2024")
DIM = Path("../data/gold/dimensions")
OUT = Path("../data/gold/facts")

OUT.mkdir(parents=True, exist_ok=True)

# Silver source
players_stats = pd.read_parquet(
    BASE / "players_stats/players_stats.parquet"
)

# Dimensions
dim_player = pd.read_parquet(DIM / "dim_player.parquet")
dim_team = pd.read_parquet(DIM / "dim_team.parquet")
dim_tournament = pd.read_parquet(DIM / "dim_tournament.parquet")
dim_agent = pd.read_parquet(DIM / "dim_agent.parquet")

mask_aggregate = players_stats["agent_name"].str.contains(",")

print(len(players_stats))
print(len(players_stats[mask_aggregate]))

players_stats = players_stats[
    ~mask_aggregate
].copy()

print(len(players_stats))

players_stats = players_stats[
    (players_stats["stage_name"] != "All Stages") &
    (players_stats["match_type_name"] != "All Match Types")
].copy()

print(len(players_stats))

fact = players_stats.merge(
    dim_player,
    on="player_name",
    how="left"
)

print(fact["player_id"].isna().sum())

fact = fact.merge(
    dim_team,
    on="team_name",
    how="left"
)

print(fact["team_id"].isna().sum())

fact = fact.merge(
    dim_agent,
    on="agent_name",
    how="left"
)

print(fact["agent_id"].isna().sum())

fact = fact.merge(
    dim_tournament,
    on=[
        "tournament_name",
        "stage_name",
        "match_type_name"
    ],
    how="left"
)

print(fact["tournament_id"].isna().sum())

len(fact.columns)

fact_player_competition_stats = fact[
    [
        # Keys
        "player_id",
        "team_id",
        "agent_id",

        "tournament_id",
        "stage_id",
        "match_type_id",

        # Volume
        "rounds_played",

        # Ratings
        "rating",
        "average_combat_score",

        # Ratios
        "kills_deaths_ratio",
        "kills_per_round",
        "assists_per_round",
        "first_kills_per_round",
        "first_deaths_per_round",

        # Percentages
        "kill_assist_trade_survive_percentage",
        "average_damage_per_round",
        "headshot_percentage",
        "clutch_success_percentage",

        # Totals
        "maximum_kills_single_map",
        "kills",
        "deaths",
        "assists",
        "first_kills",
        "first_deaths",

        # Clutches
        "clutches_won",
        "clutches_played",
    ]
].copy()


len(fact_player_competition_stats.columns)

fact_player_competition_stats.head()

out_path = OUT / "fact_player_competition_stats.parquet"

fact_player_competition_stats.to_parquet(
    out_path,
    index=False
)
